+++ src/main/java/com/buuz135/simpleclaims/commands/subcommand/party/PartyInviteCommand.java
@@ -2,6 +2,7 @@ package com.buuz135.simpleclaims.commands.subcommand.party;
 
 import com.buuz135.simpleclaims.Main;
 import com.buuz135.simpleclaims.claim.ClaimManager;
+import com.buuz135.simpleclaims.claim.party.PartyOverrides;
 import com.buuz135.simpleclaims.commands.CommandMessages;
 import com.buuz135.simpleclaims.gui.PartyInfoEditGui;
 import com.hypixel.hytale.component.Ref;
@@ -47,6 +48,10 @@ public class PartyInviteCommand extends AbstractAsyncCommand {
                         var party = ClaimManager.getInstance().getPartyFromPlayer(playerRef.getUuid());
                         if (party == null) {
                             player.sendMessage(CommandMessages.NOT_IN_A_PARTY);
+                            return;
+                        }
+                        if (!party.hasPermission(playerRef.getUuid(), PartyOverrides.PARTY_PROTECTION_INVITE_PLAYERS)) {
+                            player.sendMessage(CommandMessages.NO_PERMISSION);
                             return;
                         }
                         PlayerRef invitedPlayer = commandContext.get(this.name);